+++ backend/canisters/group/impl/src/lifecycle/post_upgrade.rs
@@ -17,10 +17,15 @@ fn post_upgrade(args: Args) {
     let memory = get_upgrades_memory();
     let reader = get_reader(&memory);
 
-    let (data, errors, logs, traces): (Data, Vec<LogEntry>, Vec<LogEntry>, Vec<LogEntry>) =
+    let (mut data, errors, logs, traces): (Data, Vec<LogEntry>, Vec<LogEntry>, Vec<LogEntry>) =
         msgpack::deserialize(reader).unwrap();
 
     canister_logger::init_with_logs(data.test_mode, errors, logs, traces);
+
+    data.chat.link_threads_to_members();
+    if data.chat.events.init_thread_messages_to_update_in_stable_memory() {
+        data.stable_memory_event_migration_complete = false;
+    }
 
     let env = init_env(data.rng_seed);
     init_state(env, data, args.wasm_version);