+++ backend/libraries/chat_events/src/chat_event_internal.rs
@@ -1,7 +1,6 @@
 use crate::{incr, MessageContentInternal};
 use serde::{Deserialize, Serialize};
 use std::cmp::max;
-use std::collections::hash_map::Entry::{Occupied, Vacant};
 use std::collections::{HashMap, HashSet};
 use std::ops::DerefMut;
 use types::{
@@ -348,11 +347,12 @@ impl From<&MembersAddedToPublicChannelInternal> for MembersAddedToDefaultChannel
 }
 
 #[derive(Serialize, Deserialize, Clone, Debug, Default)]
+#[serde(from = "ThreadSummaryInternalCombined")]
 pub struct ThreadSummaryInternal {
-    #[serde(rename = "i")]
-    pub participant_ids: Vec<UserId>,
-    #[serde(default, rename = "f")]
-    pub follower_ids: HashMap<UserId, Timestamped<bool>>,
+    #[serde(rename = "p")]
+    pub participants: Vec<UserId>,
+    #[serde(rename = "f")]
+    pub followers: HashSet<UserId>,
     #[serde(rename = "r")]
     pub reply_count: u32,
     #[serde(rename = "e")]
@@ -361,11 +361,59 @@ pub struct ThreadSummaryInternal {
     pub latest_event_timestamp: TimestampMillis,
 }
 
+#[derive(Serialize, Deserialize)]
+#[serde(untagged)]
+enum Followers {
+    Old(HashMap<UserId, Timestamped<bool>>),
+    New(HashSet<UserId>),
+}
+
+#[derive(Serialize, Deserialize)]
+pub struct ThreadSummaryInternalCombined {
+    #[serde(rename = "p", alias = "i")]
+    participants: Vec<UserId>,
+    #[serde(rename = "f")]
+    followers: Followers,
+    #[serde(rename = "r")]
+    reply_count: u32,
+    #[serde(rename = "e")]
+    latest_event_index: EventIndex,
+    #[serde(rename = "t")]
+    latest_event_timestamp: TimestampMillis,
+}
+
+impl From<ThreadSummaryInternalCombined> for ThreadSummaryInternal {
+    fn from(value: ThreadSummaryInternalCombined) -> Self {
+        let followers = match value.followers {
+            Followers::Old(map) => {
+                let mut followers: HashSet<_> = value.participants.iter().copied().collect();
+                for (user_id, following) in map {
+                    if following.value {
+                        followers.insert(user_id);
+                    } else {
+                        followers.remove(&user_id);
+                    }
+                }
+                followers
+            }
+            Followers::New(set) => set,
+        };
+
+        ThreadSummaryInternal {
+            participants: value.participants,
+            followers,
+            reply_count: value.reply_count,
+            latest_event_index: value.latest_event_index,
+            latest_event_timestamp: value.latest_event_timestamp,
+        }
+    }
+}
+
 impl ThreadSummaryInternal {
     pub fn hydrate(&self, my_user_id: Option<UserId>) -> ThreadSummary {
         ThreadSummary {
-            participant_ids: self.participant_ids.clone(),
-            followed_by_me: my_user_id.map_or(false, |u| self.follower_ids.get(&u).map_or(false, |t| t.value)),
+            participant_ids: self.participants.clone(),
+            followed_by_me: my_user_id.map_or(false, |u| self.followers.contains(&u)),
             reply_count: self.reply_count,
             latest_event_index: self.latest_event_index,
             latest_event_timestamp: self.latest_event_timestamp,
@@ -376,60 +424,25 @@ impl ThreadSummaryInternal {
         &mut self,
         sender: UserId,
         mentioned_users: &[UserId],
+        root_message_sender: UserId,
         latest_event_index: EventIndex,
         now: TimestampMillis,
     ) {
         self.latest_event_index = latest_event_index;
         self.latest_event_timestamp = now;
         self.reply_count += 1;
-        self.participant_ids.push_if_not_contains(sender);
-        self.follower_ids.remove(&sender);
+        self.participants.push_if_not_contains(sender);
+        self.followers.insert(sender);
 
         // If a user is mentioned in a thread they automatically become a follower
         for user_id in mentioned_users {
-            if !self.participant_ids.contains(user_id) {
-                self.set_follow(*user_id, now, true);
-            }
-        }
-    }
-
-    pub fn set_follow(&mut self, user_id: UserId, now: TimestampMillis, follow: bool) -> bool {
-        if self.participant_ids.contains(&user_id) {
-            return false;
+            self.followers.insert(*user_id);
         }
 
-        let new_entry = Timestamped::new(follow, now);
-        match self.follower_ids.entry(user_id) {
-            Occupied(mut e) => {
-                if e.get().value == follow {
-                    false
-                } else {
-                    e.insert(new_entry);
-                    true
-                }
-            }
-            Vacant(e) => {
-                e.insert(new_entry);
-                true
-            }
+        let is_first_message = self.reply_count == 1;
+        if is_first_message {
+            self.followers.insert(root_message_sender);
         }
-    }
-
-    pub fn participants_and_followers(&self, include_unfollowed: bool) -> Vec<UserId> {
-        self.participant_ids
-            .iter()
-            .copied()
-            .chain(
-                self.follower_ids
-                    .iter()
-                    .filter(|(_, t)| include_unfollowed || t.value)
-                    .map(|(user_id, _)| *user_id),
-            )
-            .collect()
-    }
-
-    pub fn get_follower(&self, user_id: UserId) -> Option<Timestamped<bool>> {
-        self.follower_ids.get(&user_id).cloned()
     }
 }
 
@@ -629,7 +642,7 @@ mod tests {
         TextContentInternal, ThreadSummaryInternal,
     };
     use candid::Principal;
-    use std::collections::{HashMap, HashSet};
+    use std::collections::HashSet;
     use types::{EventWrapperInternal, Reaction, Tips};
 
     #[test]
@@ -690,8 +703,8 @@ mod tests {
                 timestamp: 1,
             }),
             thread_summary: Some(ThreadSummaryInternal {
-                participant_ids: vec![principal.into()],
-                follower_ids: HashMap::new(),
+                participants: vec![principal.into()],
+                followers: HashSet::new(),
                 reply_count: 1,
                 latest_event_index: 1.into(),
                 latest_event_timestamp: 1,